update: dovetail the errors
